+++ packages/react/src/components/badge/badge.content.tsx
@@ -1,7 +1,0 @@
-import * as React from "react";
-
-export const children = (
-  <div className="dso-rich-content">
-    <p>Er zijn wijzigingsbesluiten genomen. Bekijk een besluit om de veranderingen te zien.</p>
-  </div>
-);